--- a/src/musubi_tuner/qwen_image_cache_latents.py
+++ b/src/musubi_tuner/qwen_image_cache_latents.py
@@ -84,6 +84,10 @@ def main():
 
     args = parser.parse_args()
 
+    if args.disable_cudnn_backend:
+        print("Disabling cuDNN PyTorch backend.")
+        torch.backends.cudnn.enabled = False
+
     if args.vae_dtype is not None:
         raise ValueError("VAE dtype is not supported in Qwen-Image.")