@Enyakk 2025-09-28
Even though it's not supported I am testing musubi with torch 2.8.0. However I've run into this error:
File "D:\ML\SD\trainers\musubi-new\src\musubi_tuner\qwen_image_cache_latents.py", line 87, in main if args.disable_cudnn_backend: AttributeError: 'Namespace' object has no attribute 'disable_cudnn_backend'

a relatively easy fix is:
`if hasattr(args, "disable_cudnn_backend") and args.disable_cudnn_backend:
    logger.info("Disabling cuDNN PyTorch backend.")
    torch.backends.cudnn.enabled = False
`