@xzuyn 2025-10-01
It seems it would be a good idea to add the arg to more of the scripts. qwen image edit text encoder caching is quite slow on the first run without it.

`qwen_image_cache_text_encoder_outputs.py --edit --batch_size 16`
- first run with arg: `100it [01:33,  1.06it/s]`
- second run with arg: `100it [01:33,  1.07it/s]`
- first run without arg: `100it [06:39,  4.00s/it]`
- second run without arg `100it [01:22,  1.22it/s]`